+++ eslint.config.js
@@ -1,5 +1,5 @@
 import antfu from '@antfu/eslint-config'
 
 export default antfu({
-  ignores: ['src-tauri/src/**/*'],
+  ignores: ['src-tauri/src/**/*', 'node_modules/**/*', 'src-tauri/target/**/*'],
 })